Move value signal cores to own module
